# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Flood Modeling Platform API
  version: 1.1.0
  description: >
    API contract for web-enabled HEC-RAS operations including identity,
    project/scenario management, artifact I/O, asynchronous run orchestration,
    and internal worker callbacks.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true

    RunStatus:
      type: string
      enum: [PENDING, VALIDATING, RUNNING, SUCCEEDED, FAILED, CANCELED]

    ProjectCreate:
      type: object
      required: [org_id, name]
      properties:
        org_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }

    RunCreate:
      type: object
      required: [project_id, scenario_id, input_artifact_ids]
      properties:
        project_id: { type: string, format: uuid }
        scenario_id: { type: string, format: uuid }
        input_artifact_ids:
          type: array
          items: { type: string, format: uuid }
        priority:
          type: string
          enum: [low, normal, high]
          default: normal

    WorkerProgress:
      type: object
      required: [progress_pct]
      properties:
        progress_pct:
          type: number
          minimum: 0
          maximum: 100
        log_chunk:
          type: array
          items:
            type: object
            required: [level, message]
            properties:
              level: { type: string, enum: [DEBUG, INFO, WARN, ERROR] }
              message: { type: string }

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      responses:
        '200': { description: OK }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /projects:
    get:
      tags: [Projects]
      summary: List projects
      parameters:
        - in: query
          name: org_id
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        '200': { description: OK }
    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201': { description: Created }

  /scenarios/{scenario_id}:
    get:
      tags: [Scenarios]
      summary: Get scenario details
      parameters:
        - in: path
          name: scenario_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }

  /runs:
    post:
      tags: [Runs]
      summary: Submit async run job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCreate' }
      responses:
        '202': { description: Accepted }

  /runs/{run_id}:
    get:
      tags: [Runs]
      summary: Get run status
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }

  /runs/{run_id}/logs:
    get:
      tags: [Runs]
      summary: Fetch paginated logs
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 2000, default: 200 }
      responses:
        '200': { description: OK }

  /runs/{run_id}/cancel:
    post:
      tags: [Runs]
      summary: Cancel queued/running run
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }

  /runs/{run_id}/layers:
    get:
      tags: [Results]
      summary: List map-renderable result layers
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }

  /internal/runs/{run_id}/progress:
    post:
      tags: [InternalWorkers]
      summary: Worker progress callback
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerProgress' }
      responses:
        '200': { description: OK }

  /internal/runs/{run_id}/complete:
    post:
      tags: [InternalWorkers]
      summary: Worker completion callback
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }

  /internal/runs/{run_id}/fail:
    post:
      tags: [InternalWorkers]
      summary: Worker failure callback
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }